pyproject.toml
[tool.poetry]
name = "app-python-shipdetector-onnx"
description = "A sample app leveraging the Microsoft Azure Orbital SDK to do AI-based shiped detection"
authors = ["Azure Space SDK"]
version = "1.0.0"
readme = "README.md"

[[tool.poetry.packages]]
include = "app"
from = "src"

[[tool.poetry.include]]
path = ".protos/datagenerator/planetary_computer"

[tool.poetry.dependencies]
python = ">=3.8,<3.12"
onnx = "*"
onnxruntime = "*"
Pillow = "*"
rasterio = "1.3.3"
opencv-python-headless = "^4.8.1.78"
grpcio-tools = "^1.26.0"
grpcio = "^1.26.0"
protobuf = "^3.20.1"
numpy = "1.24.4"

[tool.poetry.dependencies.microsoftazurespacefx]
path = ".wheel/microsoftazurespacefx-0.11.0-py3-none-any.whl"

[tool.poetry.group.spacefx-dev.dependencies]
pytest = "^7.2.1"
pip = "*"
setuptools = "*"
wheel = "*"
build = "*"
mypy = "*"
mypy-extensions = "*"
mypy-protobuf = "*"
debugpy = "*"

[build-system]
requires = ["poetry>=1.3.2"]
build-backend = "poetry.masonry.api"
